1.0.8: remade the GameOrderResponseDialogFragment to have only one id property once more, removing the redundant separate targetPlayerId, added ability to respond to requests in the Activity screen, changed the way we receive the result from postRequestResponse to properly throw an error to be caught and handled as an error, fixed logging tags in GameRepository, cracked the code how to update the invite Ui in a proper fashion, added filter to PlayersViewModel (DataSource) to remove the player's profile from the list of all players, when filter option is picked in GameFragment it now scrolls to show the whole filter's name, added a condition preventing you to doubly invite someone in PlayersAdapter, added a proper not-yours check for game requests in Feed and Game screens, removed unnecessary values

+++ app/src/main/java/tennis/bot/mobile/feed/activityfeed/FeedApi.kt
@@ -80,7 +80,7 @@ sealed class PostParent {
 		val opponentsCount: Int,
 		val opponentsPower: Int,
 		val date: String?,
-		val gameOrderId: Int
+		val gameOrderId: Long
 	) : PostParent()
 
 	@Serializable
